Split up ConverterTests.cs into class-specific fixtures
